Added help menu and opacity option

+++ Grid.h
@@ -4,6 +4,7 @@
 #include "Cells.h"
 #include "AABB.h"
 #include "HashTable.h"
+#include <stdint.h>
 
 typedef struct Grid
 {
@@ -12,13 +13,13 @@ typedef struct Grid
     int width;
     int height;
     Cell* cells;
-    int len;
+    size_t len;
 } Grid;
 
 Grid* makeGrid(int width, int height, double cellSize);
 Collider* makeCollider(Box hitbox, Grid* mainGrid, void* sprite);
 void freeCollider(Collider* collider);
-void insertToGrid(Grid* grid, Collider* collider, int curr_update);
-int queryBox(Grid* grid, Box box, Collider** ret_array, hashTable* table, int MAX_SIZE, int curr_update, int htable_use, int revx, int revy);
+void insertToGrid(Grid* grid, Collider* collider, uint32_t curr_update);
+int queryBox(Grid* grid, Box box, Collider** ret_array, hashTable* table, int MAX_SIZE, uint32_t curr_update, uint32_t htable_use, int revx, int revy);
 
 #endif // GRID_H_INCLUDED
