+++ HashTable.h
@@ -4,15 +4,15 @@
 
 typedef struct _hashItem
 {
-    int updateCount;
+    uint32_t updateCount;
     intptr_t value;
 } hashItem;
 
 typedef struct _hashTable
 {
     hashItem* items;
-    int len;
-    int num;
+    size_t len;
+    size_t num;
 } hashTable;
 
 void insertHashItem(hashTable* table, intptr_t value, int update);
